@@ -1,16 +1,15 @@
 module github.com/argoproj-labs/argocd-operator
 
-go 1.19
+go 1.21
 
 require (
-	github.com/argoproj/argo-cd/v2 v2.8.3
+	github.com/argoproj/argo-cd/v2 v2.10.1
 	github.com/coreos/prometheus-operator v0.40.0
-	github.com/go-logr/logr v1.2.4
-	github.com/google/go-cmp v0.5.9
+	github.com/go-logr/logr v1.3.0
+	github.com/google/go-cmp v0.6.0
 	github.com/json-iterator/go v1.1.12
-	github.com/keycloak/keycloak-operator v0.0.0-20221116085200-4b9abfb29226
 	github.com/onsi/ginkgo v1.16.5
-	github.com/onsi/gomega v1.27.6
+	github.com/onsi/gomega v1.27.10
 	github.com/openshift/api v3.9.1-0.20190916204813-cdbe64fb0c91+incompatible
 	github.com/openshift/client-go v0.0.0-20200325131901-f7baeb993edb
 	github.com/operator-framework/operator-sdk v0.18.2
