@@ -23,7 +23,7 @@ import (
 	argoprojv1alpha1 "github.com/argoproj-labs/argocd-operator/api/v1alpha1"
 	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
 	"github.com/argoproj-labs/argocd-operator/common"
-	"github.com/argoproj-labs/argocd-operator/controllers/argoutil"
+	"github.com/argoproj-labs/argocd-operator/pkg/argoutil"
 )
 
 // generateBackupKey will generate and return the backup key for the export process.
@@ -38,7 +38,7 @@ func generateBackupKey() ([]byte, error) {
 }
 
 // reconcileExport will ensure that the resources for the export process are present for the ArgoCDExport.
-func (r *ReconcileArgoCDExport) reconcileExport(cr *argoprojv1alpha1.ArgoCDExport) error {
+func (r *ArgoCDExportReconciler) reconcileExport(cr *argoprojv1alpha1.ArgoCDExport) error {
 	log.Info("reconciling export secret")
 	if err := r.reconcileExportSecret(cr); err != nil {
 		return err
@@ -59,9 +59,18 @@ func (r *ReconcileArgoCDExport) reconcileExport(cr *argoprojv1alpha1.ArgoCDExpor
 	return nil
 }
 
+// FetchStorageSecretName will return the name of the Secret to use for the export process.
+func FetchStorageSecretName(export *argoprojv1alpha1.ArgoCDExport) string {
+	name := argoutil.NameWithSuffix(export.ObjectMeta.Name, "export")
+	if export.Spec.Storage != nil && len(export.Spec.Storage.SecretName) > 0 {
+		name = export.Spec.Storage.SecretName
+	}
+	return name
+}
+
 // reconcileExportSecret will ensure that the Secret used for the export process is present.
-func (r *ReconcileArgoCDExport) reconcileExportSecret(cr *argoprojv1alpha1.ArgoCDExport) error {
-	name := argoutil.FetchStorageSecretName(cr)
+func (r *ArgoCDExportReconciler) reconcileExportSecret(cr *argoprojv1alpha1.ArgoCDExport) error {
+	name := FetchStorageSecretName(cr)
 	// Dummy CR to retrieve secret
 	a := &argoproj.ArgoCD{}
 	a.ObjectMeta = cr.ObjectMeta
@@ -96,7 +105,7 @@ func (r *ReconcileArgoCDExport) reconcileExportSecret(cr *argoprojv1alpha1.ArgoC
 }
 
 // validateExport will ensure that the given ArgoCDExport is valid.
-func (r *ReconcileArgoCDExport) validateExport(cr *argoprojv1alpha1.ArgoCDExport) error {
+func (r *ArgoCDExportReconciler) validateExport(cr *argoprojv1alpha1.ArgoCDExport) error {
 	if len(cr.Status.Phase) <= 0 {
 		cr.Status.Phase = "Pending"
 		return r.Client.Status().Update(context.TODO(), cr)
